ADD PASSWORD STRENGHT EVALUATION CLASS
Added a class to evaluate a passed in password string. The algorithm is weighted so that it returns
an exponentially larger index for inclusion of different cases, number, symbols and length.

Unit tests developed first and then code generated to make tests pass.

TO DO: Develop an overall controller class that will instantiate different classes (with required interface) dependant on input