ADD PASSWORD HANDLER CLASS TO SELECT APPROPRIATE CLASS TO RETURN BASED ON INPUT FIELD
The password handler processes the supplied input and returns the appropriate class.
Each class has the generate method since they implement the password interface.
Passing "plain" and then calling the generate method with a number generates a password
of that length with just number and characters.
Mixed includes symbols
Passing null to the handler and then calling the generate method with string returns a
password index that gives an assessment of its strength.
Unit tests set up initially and then made to pass with code as per TDD practice...
TODO - complete a README file